move presentation utils to base package and make the base view model open class
